[SOIN] Déplace les données de la visite guidée pour le tableau de bord
... afin d'éviter qu'un import depuis "EtapeHomologuer" soit caché par erreur, via le entrypoint de tableauDeBord.ts
