+++ svelte/lib/tableauDeBord/tableauDeBord.ts
@@ -1,8 +1,5 @@
 import TableauDeBord from './TableauDeBord.svelte';
-import type {
-  TableauDeBordProps,
-  NiveauSecuriteService,
-} from './tableauDeBord.d';
+import type { TableauDeBordProps } from './tableauDeBord.d';
 import RapportTeleversementServicesV2 from './televersementServices/RapportTeleversementServicesV2.svelte';
 import { mount, unmount } from 'svelte';
 
@@ -34,41 +31,6 @@ const rechargeRapportTeleversementV2 = async () => {
   rapportV2 = mount(RapportTeleversementServicesV2, {
     target: document.getElementById('rapport-televersement')!,
   });
-};
-
-export const donneesVisiteGuidee = {
-  resume: {
-    nombreServices: 1,
-    nombreServicesHomologues: 0,
-    nombreHomologationsExpirees: 0,
-  },
-  indiceCyber: 4.3,
-  services: [
-    {
-      id: 'ID-SERVICE-VISITE-GUIDEE',
-      nomService: 'MonServiceSécurisé',
-      organisationResponsable: 'ANSSI',
-      contributeurs: [],
-      statutHomologation: {
-        id: 'activee',
-        enCoursEdition: false,
-        libelle: 'Active',
-        ordre: 5,
-      },
-      nombreContributeurs: 3,
-      estProprietaire: true,
-      estAdmin: false,
-      documentsPdfDisponibles: ['annexes', 'syntheseSecurite'],
-      permissions: {
-        gestionContributeurs: true,
-      },
-      aUneSuggestionAction: false,
-      actionRecommandee: undefined,
-      niveauSecurite: 'niveau1' as NiveauSecuriteService,
-      pourcentageCompletude: 0.5,
-    },
-  ],
-  indicesCyber: [{ id: 'ID-SERVICE-VISITE-GUIDEE', indiceCyber: '4.3' }],
 };
 
 export default tdb!;
